Moved Main and CLI classes to own cli package
