+++ pom.xml
@@ -205,7 +205,7 @@
 					<archive>
 						<manifest>
 							<addClasspath>true</addClasspath>
-							<mainClass>de.nqueensfaf.Main</mainClass>
+							<mainClass>de.nqueensfaf.cli.Main</mainClass>
 						</manifest>
 					</archive>
 					<descriptorRefs>
@@ -234,7 +234,7 @@
 							<winConfig>
 								<headerType>console</headerType>
 							</winConfig>
-							<mainClass>de.nqueensfaf.Main</mainClass>
+							<mainClass>de.nqueensfaf.cli.Main</mainClass>
 						</configuration>
 					</execution>
 				</executions>
