+++ src/main/java/de/nqueensfaf/cli/Main.java
@@ -1,4 +1,4 @@
-package de.nqueensfaf;
+package de.nqueensfaf.cli;
 
 import picocli.CommandLine;
